+++ client/package.json
@@ -14,10 +14,11 @@
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/react-router-dom": "^5.3.2",
     "prettier": "^2.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.0.2",
+    "react-router-dom": "5",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.3",
     "typescript": "^4.1.2",
