+++ client/src/pages/Auth.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react';
+
+export const Auth: FC = () => {
+  return <h1>Login</h1>;
+};
