+++ client/src/pages/Home.tsx
@@ -1,0 +1,11 @@
+import { Footer } from '../components/Footer';
+import { Header } from '../components/Header';
+
+export const Home = () => {
+  return (
+    <>
+      <Header />
+      <Footer />
+    </>
+  );
+};
