--- src/App.tsx
@@ -1,14 +1,0 @@
-import './App.css';
-import { Footer } from './components/Footer';
-import { Header } from './components/Header';
-
-const App = () => {
-  return (
-    <>
-      <Header />
-      <Footer />
-    </>
-  );
-};
-
-export default App;
